+++ src/CommunicationService/Classifications/ClassificationController.List.cs
@@ -1,0 +1,14 @@
+namespace CommunicationService.Classifications;
+
+public partial class ClassificationController
+{
+    [HttpGet]
+    public async Task<IActionResult> ListClassifications(CancellationToken cancellationToken)
+    {
+        var classificationsResult = await ClassificationRepository.ListClassifications(cancellationToken);
+
+        return classificationsResult.Match(
+            item => Ok(item.Select(x => x.ToClassificationResponse())),
+            Problem);
+    }
+}
