+++ src/CommunicationService/Classifications/DataModels/Classification.Constants.cs
@@ -1,0 +1,7 @@
+namespace CommunicationService.Classifications.DataModels;
+
+public partial class Classification
+{
+    public const int MinNameLength = 3;
+    public const int MaxNameLength = 50;
+}
