+++ src/CommunicationService/MetadataTypes/Api/MetadataTypeController.GetByName.cs
@@ -1,0 +1,15 @@
+namespace CommunicationService.MetadataTypes;
+
+public partial class MetadataTypeController
+{
+    [HttpGet("ByName/{name}")]
+    public async Task<IActionResult> GetMetadataTypeByName(string name, CancellationToken cancellationToken)
+    {
+        var metadataTypeResult = await MetadataTypeRepository.GetMetadataTypeByName(name, cancellationToken);
+
+        return metadataTypeResult.Match(
+            item => Ok(item.ToMetadataTypeResponse()),
+            Problem);
+    }
+
+}
