+++ src/CommunicationService/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "ServiceSqlConnectionString": "host=172.17.0.2;database=CommunicationService;password=HLF4cb4sYdDMUm4;username=user"
+    "ServiceSqlConnectionString": ""
   }
 }
