+++ tests/CommunicationService.Test/appsettings.auto-test.json
@@ -1,8 +1,8 @@
 {
   "AutoTestDbOptions": {
-    "Host": "172.17.0.2",
-    "Username":  "user",
-    "Password": "HLF4cb4sYdDMUm4",
-    "DbPrefix": "CS_AutoTest_"
+    "Host": "",
+    "Username":  "",
+    "Password": "",
+    "DbPrefix": ""
   }
 }
